DRY up error handling in provider-files
