+++ app/packages/files/dataverse-provider-file.ts
@@ -1,20 +1,11 @@
-import { inject as service } from '@ember/service';
-
-import Intl from 'ember-intl/services/intl';
-import Toast from 'ember-toastr/services/toast';
-
 import { FileSortKey } from 'ember-osf-web/packages/files/file';
 import FileProviderModel from 'ember-osf-web/models/file-provider';
 import DataverseFile from 'ember-osf-web/packages/files/dataverse-file';
 import ProviderFile from 'ember-osf-web/packages/files/provider-file';
 import CurrentUserService from 'ember-osf-web/services/current-user';
-import captureException, { getApiErrorMessage } from 'ember-osf-web/utils/capture-exception';
 
 export default class DataverseProviderFile extends ProviderFile {
     providerHandlesVersioning = false;
-
-    @service intl!: Intl;
-    @service toast!: Toast;
 
     constructor(currentUser: CurrentUserService,providerFileModel: FileProviderModel) {
         super(currentUser, providerFileModel);
@@ -31,12 +22,7 @@ export default class DataverseProviderFile extends ProviderFile {
             this.totalFileCount = queryResult.meta.total;
             return queryResult.map(fileModel => new DataverseFile(this.currentUser, fileModel));
         } catch (e) {
-            const errorMessage = this.intl.t(
-                'osf-components.file-browser.errors.load_file_list',
-            );
-            captureException(e, { errorMessage });
-            this.toast.error(getApiErrorMessage(e), errorMessage);
-            return [];
+            return this.handleFetchError(e);
         }
     }
