+++ app/packages/files/github-provider-file.ts
@@ -1,18 +1,10 @@
-import { inject as service } from '@ember/service';
-
-import Intl from 'ember-intl/services/intl';
-import Toast from 'ember-toastr/services/toast';
 import { FileSortKey } from 'ember-osf-web/packages/files/file';
 import FileProviderModel from 'ember-osf-web/models/file-provider';
 import GithubFile from 'ember-osf-web/packages/files/github-file';
 import ProviderFile from 'ember-osf-web/packages/files/provider-file';
 import CurrentUserService from 'ember-osf-web/services/current-user';
-import captureException, { getApiErrorMessage } from 'ember-osf-web/utils/capture-exception';
 
 export default class GithubProviderFile extends ProviderFile {
-    @service intl!: Intl;
-    @service toast!: Toast;
-
     constructor(currentUser: CurrentUserService,providerFileModel: FileProviderModel) {
         super(currentUser, providerFileModel);
     }
@@ -28,12 +20,7 @@ export default class GithubProviderFile extends ProviderFile {
             this.totalFileCount = queryResult.meta.total;
             return queryResult.map(fileModel => new GithubFile(this.currentUser, fileModel));
         } catch (e) {
-            const errorMessage = this.intl.t(
-                'osf-components.file-browser.errors.load_file_list',
-            );
-            captureException(e, { errorMessage });
-            this.toast.error(getApiErrorMessage(e), errorMessage);
-            return [];
+            return this.handleFetchError(e);
         }
     }
